# openapi.yaml
openapi: 3.1.0
info:
  title: UARE Custom Backend API
  version: 1.0.0
  description: |
    Production API for the UARE enterprise platform — auth, billing, orgs,
    CAD execution, workflows, workers, analytics, patents, and ops quality.
servers:
  - url: http://localhost:8787
    description: Local development
  - url: https://api.uare.app
    description: Production
security:
  - cookieAuth: []
  - devHeaderAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: uare_session
    devHeaderAuth:
      type: apiKey
      in: header
      name: x-dev-user-id
    opsBearerToken:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
...
    Subscription:
      type: object
      properties:
        subscription_id:
          type: string
        user_id:
          type: string
        plan_id:
          type: string
        status:
          type: string
          enum: [active, past_due, canceled, trialing]
        billing_provider:
          type: string
        current_period_start:
          type: string
          format: date-time
        current_period_end:
          type: string
          format: date-time
          nullable: true
        metadata_json:
          type: object
    CheckoutSession:
      type: object
      properties:
        checkout_session_id:
          type: string
        user_id:
          type: string
        plan_id:
          type: string
        status:
          type: string
        checkout_url:
          type: string
          nullable: true
        provider:
          type: string
    Invoice:
      type: object
      properties:
        invoice_id:
          type: string
        number:
          type: string
          nullable: true
        status:
          type: string
        amount_due:
          type: integer
        amount_paid:
          type: integer
        currency:
          type: string
        created:
          type: integer
        due_date:
          type: integer
          nullable: true
        hosted_invoice_url:
          type: string
          nullable: true
        invoice_pdf:
          type: string
          nullable: true
        period_start:
          type: integer
        period_end:
          type: integer
...
paths:
...
  /billing/summary:
    get:
      tags: [billing]
      summary: Get full billing summary including entitlements and usage
      parameters:
        - in: query
          name: org_id
          schema:
            type: string
      responses:
        '200':
          description: Billing summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  user:
                    type: object
                  subscription:
                    $ref: '#/components/schemas/Subscription'
                  usage:
                    type: object
                  entitlements:
                    type: object
  /billing/checkout-session:
    post:
      tags: [billing]
      summary: Create a checkout session to upgrade or start a subscription
      parameters:
        - in: header
          name: Idempotency-Key
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                plan_id:
                  type: string
                  enum: [free, pro, studio, enterprise]
                  default: pro
                org_id:
                  type: string
      responses:
        '201':
          description: Checkout session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  session:
                    $ref: '#/components/schemas/CheckoutSession'
                  provider:
                    type: string
  /billing/portal-session:
    post:
      tags: [billing]
      summary: Create a Stripe customer portal session (manage subscription/invoices)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                return_url:
                  type: string
                  format: uri
                org_id:
                  type: string
      responses:
        '201':
          description: Portal session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  portal_url:
                    type: string
                  portal_session_id:
                    type: string
                    nullable: true
                  stub:
                    type: boolean
        '422':
          description: No Stripe customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /billing/invoices:
    get:
      tags: [billing]
      summary: List recent invoices from Stripe
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            maximum: 25
            default: 10
      responses:
        '200':
          description: Invoice list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  invoices:
                    type: array
                    items:
                      $ref: '#/components/schemas/Invoice'
                  provider:
                    type: string
  /billing/webhook/stripe:
    post:
      tags: [billing]
      summary: Stripe webhook receiver
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Event processed
        '400':
          description: Invalid signature

  # ── Organizations ─────────────────────────────────────────────────────────
...